Add my_jss

+++ JssProject/main/views.py
@@ -2,12 +2,23 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .forms import JssForm
 from .models import Jasoseol
 from django.http import Http404
+from django.core.exceptions import PermissionDenied
+from django.contrib.auth.decorators import login_required
 
 def index(request):
     all_jss = Jasoseol.objects.all()
     return render(request, 'index.html', {'all_jss':all_jss})
 
+def my_index(request):
+    my_jss = Jasoseol.objects.filter(author=request.user)
+
+    return render(request, 'index.html', {'all_jss':my_jss})
+
+@login_required(login_url='/login/') # decorators
 def create(request):
+    #if not request.user.is_authenticated:
+        #return redirect('login')
+
     if request.method == "POST":
         filled_form = JssForm(request.POST)
         if filled_form.is_valid(): # 유효성 검증
@@ -18,6 +29,7 @@ def create(request):
     jss_form = JssForm()
     return render(request, 'create.html', {'jss_form':jss_form})
 
+@login_required(login_url='/login/')
 def detail(request, jss_id):
     # try:
     #    my_jss = Jasoseol.objects.get(pk=jss_id)
@@ -30,9 +42,11 @@ def detail(request, jss_id):
 
 def delete(request, jss_id):
     my_jss = Jasoseol.objects.get(pk=jss_id)
-    my_jss.delete()
+    if request.user == my_jss.author:
+        my_jss.delete()
+        return redirect('index')
 
-    return redirect('index')
+    raise PermissionDenied
 
 def update(request, jss_id):
     my_jss = Jasoseol.objects.get(pk=jss_id)
